refactor: improve type safety and code quality across codebase
Enhance type annotations, fix type checker errors, improve code formatting,
and strengthen test infrastructure for better maintainability.

Changes:
- Inference: Update return types to support union types (Tensor/tuple/dict)
  for models with multiple outputs, add type narrowing in evaluate.py
- Fingerprinting: Add per-image intensity statistics (mean, std, percentiles)
  to ImageProperties for better dataset characterization
- Testing: Add session-scoped nnUNet environment variable fixture to ensure
  proper test isolation and create preprocessed directory structure in mocks
- Cropping: Replace type ignore with explicit np.asarray() for better type safety
- Plotting: Add clarifying comments for matplotlib style side-effect imports
- Dataset helpers: Apply consistent code formatting (line breaks, whitespace)
- Tests: Add isinstance checks for type narrowing and update mocks with new
  fingerprint fields
- Config: Add pytest autouse fixture to ruff ignore list

All changes maintain backward compatibility while improving code quality
and type safety for static analysis tools.

+++ src/planning/planner/create.py
@@ -202,7 +202,7 @@ def create_experiment_plan(
     )
 
     # Convert fingerprint numpy arrays to Python tuples for serialization
-    median_shape_py = tuple(float(x) for x in fingerprint.median_shape)
+    median_shape_py = tuple(int(x) for x in fingerprint.median_shape)
     median_spacing_py = tuple(float(x) for x in fingerprint.median_spacing)
 
     plan = ExperimentPlan(
